fix(023): move logIngestionAttempt to server-only module
logIngestionAttempt was exported from a "use server" module, making it
callable as a Server Action without authentication. Move it to
src/lib/ingestion-logger.ts with "server-only" import so it can only
be called from server-side code.
