+++ src/actions/ingestion-log.ts
@@ -21,36 +21,6 @@ export interface IngestionLogRow {
   createdAt: string;
 }
 
-interface LogIngestionParams {
-  filename?: string | null;
-  vendor?: string | null;
-  invoiceNumber?: string | null;
-  invoiceDate?: string | null;
-  amountCents?: number | null;
-  outcome: "success" | "failed";
-  errorMessage?: string | null;
-  channel: "manual" | "api" | "bulk";
-  blobPathname?: string | null;
-  linkedInvoiceId?: number | null;
-  uploadedBy?: number | null;
-}
-
-export async function logIngestionAttempt(params: LogIngestionParams) {
-  await db.insert(ingestionLog).values({
-    filename: params.filename ?? null,
-    vendor: params.vendor ?? null,
-    invoiceNumber: params.invoiceNumber ?? null,
-    invoiceDate: params.invoiceDate ?? null,
-    amountCents: params.amountCents ?? null,
-    outcome: params.outcome,
-    errorMessage: params.errorMessage ?? null,
-    channel: params.channel,
-    blobPathname: params.blobPathname ?? null,
-    linkedInvoiceId: params.linkedInvoiceId ?? null,
-    uploadedBy: params.uploadedBy ?? null,
-  });
-}
-
 export async function getIngestionHistory(): Promise<
   { success: true; data: IngestionLogRow[] } | { success: false; error: string }
 > {
